requirements.txt
requests
pandas
discord.py
Pillow
pymongo[srv]
redis
